+++ crates/utils-sync/src/lib.rs
@@ -5,12 +5,14 @@ extern crate alloc;
 #[cfg(any(feature = "std", test))]
 extern crate std;
 
+pub mod once_lock;
 pub mod racy_lock;
 pub mod rw_lock;
 
 #[cfg(feature = "std")]
 pub use std::sync::LazyLock;
 
+pub use once_lock::OnceLockCompat;
 #[cfg(feature = "std")]
 pub use parking_lot::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 #[cfg(not(feature = "std"))]